+++ Scenes/ChatMenu/message_edit.gd
@@ -4,14 +4,14 @@ class_name TransmissionEdit
 @onready var autocomplete_list: ItemList = $PanelContainer/MarginContainer/ItemList
 @onready var auto_list_panel: PanelContainer = $PanelContainer
 #CodeEdit hates trying to autocomplete things without spaces, so use a nested one which contains only the word to try
-@onready var autocomplete_finder: CodeEdit = $AutocompleteFinder
+#@onready var autocomplete_finder: CodeEdit = $AutocompleteFinder
 
 const DELIMITER_CHARACTER = "\u001f"
 
 func _ready():
 	Main.instance.reload_settings.connect(_set_popup_size.call_deferred)
 	auto_list_panel.hide()
-	autocomplete_finder.hide()
+	#autocomplete_finder.hide()
 	syntax_highlighter = TransmissionHighlighter.new()
 
 signal submit_text(message: String)
@@ -29,23 +29,14 @@ func _request_code_completion(force: bool) -> void:
 	# TODO: make this better
 	# dont add every word?
 	# but only if it makes it lag
-	autocomplete_finder.text = word_under_caret
-	autocomplete_finder.set_caret_column(word_under_caret.length() + 1)
-	for word: String in (DictionaryHandler.word_names as Array[String]):
-		word = AutocompleteManager.encode_special_chars(word)
-		#print(word)
-		#print(AutocompleteManager.decode_special_chars(word))
-		autocomplete_finder.add_code_completion_option(CodeEdit.KIND_PLAIN_TEXT, word, word)
-	autocomplete_finder.update_code_completion_options(true)
-
-	var options: Array[Dictionary] = autocomplete_finder.get_code_completion_options()
+	var options: Array = AutocompleteManager.get_autocomplete_options(word_under_caret, DictionaryHandler.word_names)
 	if options.is_empty():
 		#print("NO OPTIONS FOR ", word_under_caret)
 		auto_list_panel.hide()
 		return
 	#print(options)
 	
-	if options.size() == 1 and options[0]["display_text"] == word_under_caret:
+	if options.size() == 1 and options[0] == word_under_caret:
 		auto_list_panel.hide()
 		var bounds: Array = get_signal_bounds_under_caret()
 		var line: String = get_line(get_caret_line())
@@ -54,7 +45,6 @@ func _request_code_completion(force: bool) -> void:
 				return
 		insert_text(DELIMITER_CHARACTER, get_caret_line(), bounds[1])
 		return
-	autocomplete_finder.cancel_code_completion()
 	
 	_show_custom_popup(options)
 
@@ -85,10 +75,10 @@ func get_signal_under_caret() -> String:
 	var bounds = get_signal_bounds_under_caret()
 	return line_text.substr(bounds[0], bounds[1] - bounds[0])
 
-func _show_custom_popup(options: Array[Dictionary]):
+func _show_custom_popup(options: Array):
 	autocomplete_list.clear()
 	for option in options:
-		autocomplete_list.add_item(AutocompleteManager.decode_special_chars(option["display_text"]))
+		autocomplete_list.add_item(AutocompleteManager.decode_special_chars(option))
 	autocomplete_list.select(0)
 	
 	var word = get_signal_under_caret()